bert-moderator/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
discord
python-dotenv
torch==1.5.0+cpu
torchvision==0.6.0+cpu
detoxify
tensorflow-cpu==1.15
tensorflow-hub==0.7.0
pillow